+++ src/main/kotlin/com/wutsi/codegen/kotlin/sdk/SdkModelCodeGenerator.kt
@@ -1,6 +1,5 @@
 package com.wutsi.codegen.kotlin.sdk
 
-import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.FileSpec
 import com.squareup.kotlinpoet.FunSpec
 import com.squareup.kotlinpoet.KModifier
@@ -16,13 +15,6 @@ import com.wutsi.codegen.model.Type
 import io.swagger.v3.oas.models.OpenAPI
 import java.time.LocalDate
 import java.time.OffsetDateTime
-import javax.validation.constraints.Max
-import javax.validation.constraints.Min
-import javax.validation.constraints.NotBlank
-import javax.validation.constraints.NotEmpty
-import javax.validation.constraints.NotNull
-import javax.validation.constraints.Pattern
-import javax.validation.constraints.Size
 
 class SdkModelCodeGenerator(private val mapper: KotlinMapper) : AbstractKotlinCodeGenerator() {
     override fun generate(openAPI: OpenAPI, context: Context) {
@@ -30,7 +22,7 @@ class SdkModelCodeGenerator(private val mapper: KotlinMapper) : AbstractKotlinCo
         models.forEach { generateModel(it, context) }
     }
 
-    private fun loadModels(spec: OpenAPI, context: Context): List<Type> {
+    fun loadModels(spec: OpenAPI, context: Context): List<Type> {
         // Load
         spec.components.schemas.map { mapper.toType(it.key, it.value) }
 
@@ -75,7 +67,7 @@ class SdkModelCodeGenerator(private val mapper: KotlinMapper) : AbstractKotlinCo
 
     fun toParameterSpec(field: Field): ParameterSpec {
         val builder = ParameterSpec.builder(field.name, field.type.asTypeName().copy(field.nullable))
-            .addAnnotations(toAnnotationSpecs(field))
+            .addAnnotations(toValidationAnnotationSpecs(field))
 
         val default = defaultValue(field)
         if (default != null)
@@ -109,67 +101,5 @@ class SdkModelCodeGenerator(private val mapper: KotlinMapper) : AbstractKotlinCo
                 else -> return null
             }
         }
-    }
-
-    fun toAnnotationSpecs(field: Field): List<AnnotationSpec> {
-        val annotations = mutableListOf<AnnotationSpec>()
-        if (field.required) {
-            if (field.type == String::class) {
-                annotations.add(
-                    AnnotationSpec.builder(NotBlank::class.java)
-                        .useSiteTarget(AnnotationSpec.UseSiteTarget.GET)
-                        .build()
-                )
-            } else {
-                annotations.add(
-                    AnnotationSpec.builder(NotNull::class.java)
-                        .useSiteTarget(AnnotationSpec.UseSiteTarget.GET)
-                        .build()
-                )
-                if (field.type == List::class) {
-                    annotations.add(
-                        AnnotationSpec.builder(NotEmpty::class.java)
-                            .useSiteTarget(AnnotationSpec.UseSiteTarget.GET)
-                            .build()
-                    )
-                }
-            }
-        }
-        if (field.min != null) {
-            annotations.add(
-                AnnotationSpec.builder(Min::class.java)
-                    .useSiteTarget(AnnotationSpec.UseSiteTarget.GET)
-                    .addMember(field.min.toString())
-                    .build()
-            )
-        }
-        if (field.max != null) {
-            annotations.add(
-                AnnotationSpec.builder(Max::class.java)
-                    .useSiteTarget(AnnotationSpec.UseSiteTarget.GET)
-                    .addMember(field.max.toString())
-                    .build()
-            )
-        }
-        if (field.minLength != null || field.maxLength != null) {
-            val builder = AnnotationSpec.builder(Size::class.java)
-                .useSiteTarget(AnnotationSpec.UseSiteTarget.GET)
-
-            if (field.minLength != null)
-                builder.addMember("min=" + field.minLength.toString())
-            if (field.maxLength != null)
-                builder.addMember("max=" + field.maxLength.toString())
-
-            annotations.add(builder.build())
-        }
-        if (field.pattern != null) {
-            annotations.add(
-                AnnotationSpec.builder(Pattern::class.java)
-                    .useSiteTarget(AnnotationSpec.UseSiteTarget.GET)
-                    .addMember("\"${field.pattern}\"")
-                    .build()
-            )
-        }
-        return annotations
     }
 }