+++ src/test/kotlin/com/wutsi/codegen/kotlin/sdk/SdkModelCodeGeneratorTest.kt
@@ -150,7 +150,7 @@ internal class SdkModelCodeGeneratorTest {
 
     @Test
     fun `generate`() {
-        val yaml = IOUtils.toString(SdkCodeGenerator::class.java.getResourceAsStream("/api.yaml"))
+        val yaml = IOUtils.toString(SdkCodeGenerator::class.java.getResourceAsStream("/api.yaml"), "utf-8")
         codegen.generate(
             openAPI = OpenAPIV3Parser().readContents(yaml).openAPI,
             context = context